pyproject.toml
[tool.poetry]
name = "telegram_bot"
version = "0.1.0"
description = ""
authors = ["mehonator <mehonator@yandex.ru>"]

[tool.poetry.dependencies]
python = "3.9.10"
python-telegram-bot = "^13.10"
python-dotenv = "^0.19.2"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
